refactor helper functions to fix import cycle
